address coderabbit: audio sink cleanup, demod count, CI pipewire
- Audio sink stop() always cleans up if handles exist, regardless of
  running flag (fixes race between worker exit and stop call)
- DSP controller surfaces SinkError::Disconnected/NotRunning to UI
  as error toast instead of silently debug-logging
- DEMOD_MODE_COUNT derived from demod_selector::DEMOD_LABELS.len()
  instead of hardcoded 8
- Feature-gate PipeWire behind optional `pipewire` feature with stub
  fallback for CI/non-Linux builds
- CI installs libpipewire-0.3-dev for PipeWire compilation
