+++ .github/workflows/ci.yml
@@ -28,7 +28,8 @@ jobs:
         run: |
           sudo apt-get update
           sudo apt-get install -y build-essential pkg-config \
-            libgtk-4-dev libadwaita-1-dev libusb-1.0-0-dev
+            libgtk-4-dev libadwaita-1-dev libusb-1.0-0-dev \
+            libpipewire-0.3-dev
       - name: Install cargo tools
         run: cargo install --locked cargo-deny cargo-audit
