+++ Cargo.toml
@@ -17,7 +17,7 @@ sdr-rtlsdr.workspace = true
 sdr-source-rtlsdr.workspace = true
 sdr-source-network.workspace = true
 sdr-source-file.workspace = true
-sdr-sink-audio.workspace = true
+sdr-sink-audio = { workspace = true, features = ["pipewire"] }
 sdr-sink-network.workspace = true
 tracing.workspace = true
 tracing-subscriber = { workspace = true, features = ["env-filter"] }
