+++ crates/sdr-sink-audio/Cargo.toml
@@ -6,15 +6,17 @@ edition.workspace = true
 license.workspace = true
 repository.workspace = true
 
+[features]
+default = []
+pipewire = ["dep:pipewire"]
+
 [dependencies]
 sdr-types.workspace = true
 sdr-pipeline.workspace = true
 sdr-config.workspace = true
 thiserror.workspace = true
 tracing.workspace = true
-
-[target.'cfg(target_os = "linux")'.dependencies]
-pipewire.workspace = true
+pipewire = { workspace = true, optional = true }
 
 [lints]
 workspace = true
